increase output buffer size to 2^17
With my current font, Terminus, I can view up to 319 * 84 chars. If
these are 3 byte UTF8 characters, then this at minimum takes in 80388
bytes. Since there are countless other color codes, let's round up to
the next power of 2 minus 1 so it fits nicely in memory.

Doing this only one draw call is needed, and any flickering or partial
draws is not VIM's fault, but rather the terminal or networking
conditions.

This apparently seemed a bit controversial and no further action was
taken, but in the era of multi GB machines, I think we can spare a few
more KBs of RAM.

original discussion here:
https://groups.google.com/forum/#!searchin/vim_dev/increase$20output$20buffer$20size/vim_dev/i8X2oIVflI0/KVXF-LzmUA8J